move main inference class
